+++ .gitignore
@@ -1,12 +1,12 @@
 # Maven projects
-/cim1-import/target/
+/cim1-converter/target/
 /cim1-model/target/
 /distribution-entsoe/target
 /entsoe-case-repository/target
 /entsoe-util/target/
 /network-merge/target/
 /target
-/ucte-import/target/
+/ucte-converter/target/
 /ucte-network/target/
 
 # IntelliJ
