+++ cim1-converter/src/main/java/eu/itesla_project/cim1/converter/CIM1Exception.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2016, All partners of the iTesla project (http://www.itesla-project.eu/consortium)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package eu.itesla_project.cim1.converter;
+
+/**
+ *
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ */
+public class CIM1Exception extends RuntimeException {
+
+    public CIM1Exception(String message) {
+        super(message);
+    }
+
+    public CIM1Exception(Throwable cause) {
+        super(cause);
+    }
+
+}
