+++ cim1-converter/src/main/resources/icons/entsoe16x16.png
