+++ pom.xml
@@ -20,13 +20,13 @@
     <url>http://www.itesla-project.eu/</url>
 
     <modules>
+        <module>cim1-converter</module>
         <module>cim1-model</module>
-        <module>cim1-import</module>
         <module>distribution-entsoe</module>
         <module>entsoe-case-repository</module>
         <module>entsoe-util</module>
         <module>network-merge</module>
-        <module>ucte-import</module>
+        <module>ucte-converter</module>
         <module>ucte-network</module>
     </modules>
