+++ ucte-converter/src/main/resources/icons/ucte16x16.png
