+++ ucte-converter/src/test/java/eu/itesla_project/ucte/converter/CoverageTest.java
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) 2017, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package eu.itesla_project.ucte.converter;
+
+import org.junit.Test;
+
+/**
+ * @author Mathieu Bague <mathieu.bague at rte-france.com>
+ */
+public class CoverageTest {
+
+    @Test
+    public void testNothing() {
+    }
+}
