--- ucte-import/src/test/java/eu/itesla_project/iidm/import_/ucte/CoverageTest.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2017, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package eu.itesla_project.iidm.import_.ucte;
-
-import org.junit.Test;
-
-/**
- * @author Mathieu Bague <mathieu.bague at rte-france.com>
- */
-public class CoverageTest {
-
-    @Test
-    public void testNothing() {
-    }
-}
